% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenTensorSamples.R
\name{covTensor}
\alias{covTensor}
\title{Performs 4th order Eigentensor analysis on m-number of n x n matrices, sampled SAMPLES times}
\usage{
covTensor(Gs)
}
\arguments{
\item{Gs}{an array of the order n x n x m x SAMPLES}
}
\value{
returns a list with summary statistics of the eigenTensor with 6 slots
$tensor.summary	Summary of the covariance tensor for the posterior mean S From left-to-right the columns of $tensor.summary are the eigenvalues of the tensor (S.eigval), the eigenvalues of eigentensors (eT.val), and the eigenvectors of eigentensors. Eigenvectors of eigentensors are stored as rows with the trait loadings identified by the column names. The eigenvectors of eigentensors are ordered in terms of the absolute value of eT.val
$av.S	Posterior mean S
$eTmat	Eigentensors of the posterior mean S. The rows and columns of $eTmat identify the elements of an eigentensor. The third dimension of $eTmat identifies the eigentensor.
$avG.coord	Coordinates of the posterior mean Gs in the space of the eigentensors of the posterior mean S.
$S	S for each sample of the set of G
$S.val	Variance of the eigentensors for each sample of the second-order representation of the fourth-order tensor
$G.coord	Coordinates of each sample of G in the space of the eigentensors of the posterior mean S.
}
\description{
Performs 4th order Eigentensor analysis on m-number of n x n matrices, sampled SAMPLES times
}
\details{
Eigentensor analysis originally published in Hine et al. (2009) to calculate the fourth-order covariance tensor of a set of second-order G matrices.
}
\examples{
covTensor(Gs)
}
