% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfSibPed.R
\name{halfSibMaker}
\alias{halfSibMaker}
\title{Simulates a Half-sib pedigree}
\usage{
halfSibMaker(s, dps, opd)
}
\arguments{
\item{s}{numeric(1), number of sires to simulate}

\item{dps}{numeric(1), number of dams mated to each sire}

\item{opd}{numeric(1), number of offspring produced by each family (dam)}
}
\value{
Returns a 3 column dataframe in order of ID,sire,dam
}
\description{
Generate random Half sib Pedigree given the number of sires, dams mated to each sire and, offspring per dam
}
\examples{
s=100;dps=4;opd=4
halfSibMaker(s, dps, opd)
}
