% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWombatDataMultivar.R
\name{makeMultivariateData}
\alias{makeMultivariateData}
\title{Makes stacked dataset for multivariate trait analysis}
\usage{
makeMultivariateData(
  idVec,
  randomEffectData = NULL,
  fixedCovVec = NULL,
  fixedEffectVec = NULL,
  dataFileName,
  dir
)
}
\arguments{
\item{idVec}{vector of ids}

\item{randomEffectData}{dataframe (or matrix) of random effect data}

\item{fixedCovVec}{numeric vector, fixed covariate variable included in the model}

\item{fixedEffectVec}{vector, fixed effect variable included in the model}

\item{dataFileName}{string, name of fixed effect variable to be included in the model}

\item{dir}{string, directory where parameter file should be output}
}
\value{
outputs multivariate data file for use in wombat run
}
\description{
Makes a stacked multivariate datafile for use in wombat. In particular this function creates a stacked datafile for multiple traits at once for multivariate analysis.
}
\examples{
dir <- getwd()
idVec <- seq(1000,1002)
randomEffectData <- matrix(rnorm(9,1,1),3,3)
fixedCovVec <- rnorm(3,5,1)
fixedEffectVec <- rnorm(3,10,1)
dataFileName <- "new"
makeMultivariateData(idVec,randomEffectData,fixedCovVec,fixedEffectVec,dataFileName,dir)
}
