% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivarParamMaker.R
\name{makeMultivariateParameter}
\alias{makeMultivariateParameter}
\title{Make multivariate parameter file}
\usage{
makeMultivariateParameter(
  randomEffectData,
  fixedCovName = NULL,
  fixedEffectName = NULL,
  pedName,
  dataFileName,
  rank = ncol(randomEffectData),
  dir
)
}
\arguments{
\item{randomEffectData}{dataframe (or matrix) of random effect data}

\item{fixedCovName}{string, name of fixed effect covariate}

\item{fixedEffectName}{string, name of fixed effect variable to be included in the model}

\item{pedName}{string, with the name of the pedigree file (extension excluded)}

\item{dataFileName}{string, name of fixed effect variable to be included in the model}

\item{rank}{integer, rank of the model to be run (wombat accommodates for reduced rank models if desirable)}

\item{dir}{string, directory where parameter file should be output}
}
\value{
Returns a stacked parameter file that, when accompanied with formatted data and pedigree files, will run the desired mixed model.
}
\description{
Makes a multivariate parameter file for use with the wombat program.
}
\examples{
randomEffectData<- matrix(rnorm(9,1,1),3,3)
fixedCovName <- "fc"
fixedEffectName <- "fe"
pedName <- "pedpedped"
dataFileName <- "data"
rank <- 1
dir<-getwd()
makeMultivariateParameter(randomEffectData,fixedCovName ,fixedEffectName, pedName, dataFileName, rank, dir)
}
