% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createuniD.R
\name{makeUnivariateData}
\alias{makeUnivariateData}
\title{Makes univariate data file}
\usage{
makeUnivariateData(
  idVec,
  randomEffectVec = NULL,
  fixedCovVec = NULL,
  fixedEffectVec = NULL,
  dataFileName,
  dir
)
}
\arguments{
\item{idVec}{vector of ids}

\item{randomEffectVec}{numeric vector of trait values for individuals (should be in the same order as idvec)}

\item{fixedCovVec}{numeric vector of fixed covariate for individuals (should be in the same order as idvec)}

\item{fixedEffectVec}{vector of fixed effect for individuals (should be in the same order as idvec)}

\item{dataFileName}{string, name of datafile}

\item{dir}{string, directory where .d file will be written}
}
\value{
Makes a univariate datafile for wombat
}
\description{
Generates a univariate data file formatted for wombat
}
\examples{
idVec<-c(1,2,3)
randomEffectVec<-c(4,5,6)
fixedCovVec<-c(7,8,9)
fixedEffectVec<-c(10,11,12)
dataFileName<-"new"
dir<-getwd()
makeUnivariateData(idVec,randomEffectVec,fixedCovVec, fixedEffectVec, dataFileName, dir)
}
