% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUniParam.R
\name{makeUnivariateParameter}
\alias{makeUnivariateParameter}
\title{Makes a univariate parameter file}
\usage{
makeUnivariateParameter(
  randomEffect,
  fixedCovName,
  fixedEffectName,
  datafileName,
  pedName,
  startingVariance,
  dir
)
}
\arguments{
\item{randomEffect}{numeric vector, of the random variable to be included in the model}

\item{fixedCovName}{character(1)string, name of fixed effect covariate}

\item{fixedEffectName}{character(1)string, name of fixed effect}

\item{datafileName}{character(1)string, name of the datafile associated with this wombat model}

\item{pedName}{character(1)string, with the name of the pedigree file (extension excluded)}

\item{startingVariance}{string, specifying the starting variance from which wombat will employ hill climbing algorithms to improve}

\item{dir}{character(1) string, specifying the dir where the parameter file should be output}
}
\value{
Makes parameter file for use with wombat software in the form of a .par file.
}
\description{
Function to make a parameter file for easy use in the Wombat quantitative genetics software
}
\examples{
randomEffect <- TRUE
fixedCovName <- "fc"
fixedEffectName <- "fe"
datafileName <- "data"
pedName <- "ped"
startingVariance <- 10
dir <- getwd()
makeUnivariateParameter(randomEffect, fixedCovName, fixedEffectName, datafileName, pedName, startingVariance, dir)
}
