% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenSim.R
\name{phenSim}
\alias{phenSim}
\title{Simulates phenotypes for individuals in a pedigree}
\usage{
phenSim(G, E, ped)
}
\arguments{
\item{G}{N dimenstional covariance matrix representing additive genetic variance (G)}

\item{E}{N dimenstional covariance matrix representing Environmental variance (E)}

\item{ped}{pedigree (Nx3 dataframe), where columns are in the order of (and labled) ID,sire,dam. Founders should have parentage of NA}
}
\value{
Returns a dataframe of the size NxM where N is the number of individuals in the pedigree and M is the number of traits considered.
}
\description{
Simulates random phenotypes for individuals in a pedigree given a pedigree and G and E matrices. Generates breeding values described in the methods of Morrissey et al(2007)
where ai=mi +C(G/2) + N
where ai is the vector of breeding values of the focal individual, mi is the vector of mid-parent values,
C(G/2) is the Cholesky decomposition of the matrix describing the segregational variance (i.e. G is the variance–covariance matrix),
and N is a vector of standard random normal deviates.
}
\examples{
#phens <- phenSim(G, E, ped)
G <- cov(matrix(sample(1:20, 100, replace = TRUE), ncol = matrixNumber))
E <- cov(matrix(sample(1:20, 100, replace = TRUE), ncol = matrixNumber))
s=100;dps=4;opd=4
ped <- halfSibMaker(s,dps,opd)
phens <- phenSim(G,E,ped)
}
