% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wombatRunner.R
\name{wombatRunner}
\alias{wombatRunner}
\title{Runs basic wombat run}
\usage{
wombatRunner(filePath)
}
\arguments{
\item{filePath}{character(1) string, specifying the filepath leading to the parameter file}
}
\value{
returns wombat to standard output in the working directory
}
\description{
Running wombat given the filepath of a parameter (.par) file
}
\details{
This function allows simple runs of wombat with the most limited options.
}
\examples{
filePath <- "A:/wombatGithub/univariatePackages/Uni.par"
wombatRunner(filePath)
}
